Added Iron Door functionality
door.open and door.close now affect iron doors
It's theoretically possible to record the door opener, but I have
decided it's far too inelegant a solution for such an impractical
feature anyway. It's easy to work around for the scripter by recording
the door opener by other means.
A delay has been instituted in order to make redstone updating work as
intended. Everything is updated by the end of a server tick anyway, but
the functionality should be reusable for other script functions as well.